refactor: split content index queries
