+++ src/codex_usage_tracker/store/content_index.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import hashlib
 import json
 import os
-import re
 import sqlite3
 from collections.abc import Callable, Iterable, Mapping
 from concurrent.futures import ThreadPoolExecutor, as_completed
@@ -26,7 +25,18 @@ from codex_usage_tracker.store.content_index_events import (
     PendingToolCall,
     extract_pending_local_events,
 )
-from codex_usage_tracker.store.query_sql import usage_where_clause
+from codex_usage_tracker.store.content_search import (
+    ContentSearchResult as ContentSearchResult,
+)
+from codex_usage_tracker.store.content_search import (
+    search_content_fragments as search_content_fragments,
+)
+from codex_usage_tracker.store.content_trace import (
+    ContentTraceResult as ContentTraceResult,
+)
+from codex_usage_tracker.store.content_trace import (
+    trace_thread_content as trace_thread_content,
+)
 
 MAX_FRAGMENT_CHARS = 4000
 DEFAULT_SEARCH_SNIPPET_CHARS = 800
@@ -93,23 +103,6 @@ class _ExtractedContentRows:
     fragment_rows: list[dict[str, object]]
     event_rows: PendingEventRows
     parse_warnings: int = 0
-
-
-@dataclass(frozen=True)
-class ContentSearchResult:
-    """Content-index search result rows and paging metadata."""
-
-    rows: list[dict[str, object]]
-    total_matched_rows: int
-    search_mode: str
-
-
-@dataclass(frozen=True)
-class ContentTraceResult:
-    """Thread/session trace result rows and paging metadata."""
-
-    calls: list[dict[str, object]]
-    total_matched_calls: int
 
 
 def index_content_for_source_files(
@@ -335,579 +328,6 @@ def delete_content_index_rows_for_source_files(
         )
     if sync_fts:
         _rebuild_content_fts(conn)
-
-
-def search_content_fragments(
-    conn: sqlite3.Connection,
-    *,
-    query: str,
-    since: str | None = None,
-    until: str | None = None,
-    model: str | None = None,
-    effort: str | None = None,
-    thread: str | None = None,
-    include_archived: bool = False,
-    limit: int | None = 20,
-    offset: int = 0,
-    max_snippet_chars: int | None = DEFAULT_SEARCH_SNIPPET_CHARS,
-) -> ContentSearchResult:
-    """Search the local content index and return bounded local snippets."""
-
-    normalized_query = query.strip()
-    if not normalized_query:
-        raise ValueError("query is required")
-    normalized_limit = _normalize_search_limit(limit)
-    normalized_offset = _normalize_search_offset(offset)
-    try:
-        return _search_content_fts(
-            conn,
-            query=normalized_query,
-            since=since,
-            until=until,
-            model=model,
-            effort=effort,
-            thread=thread,
-            include_archived=include_archived,
-            limit=normalized_limit,
-            offset=normalized_offset,
-            max_snippet_chars=max_snippet_chars,
-        )
-    except sqlite3.DatabaseError:
-        return _search_content_like(
-            conn,
-            query=normalized_query,
-            since=since,
-            until=until,
-            model=model,
-            effort=effort,
-            thread=thread,
-            include_archived=include_archived,
-            limit=normalized_limit,
-            offset=normalized_offset,
-            max_snippet_chars=max_snippet_chars,
-        )
-
-
-def trace_thread_content(
-    conn: sqlite3.Connection,
-    *,
-    thread: str | None = None,
-    thread_key: str | None = None,
-    session_id: str | None = None,
-    record_id: str | None = None,
-    since: str | None = None,
-    until: str | None = None,
-    include_archived: bool = False,
-    limit: int | None = 100,
-    offset: int = 0,
-    max_snippet_chars: int | None = DEFAULT_SEARCH_SNIPPET_CHARS,
-) -> ContentTraceResult:
-    """Return a paged aggregate call trace with attached content fragments."""
-
-    where_clause, params = _trace_where_clause(
-        thread=thread,
-        thread_key=thread_key,
-        session_id=session_id,
-        record_id=record_id,
-        since=since,
-        until=until,
-        include_archived=include_archived,
-    )
-    normalized_limit = _normalize_search_limit(limit)
-    normalized_offset = _normalize_search_offset(offset)
-    count_row = conn.execute(
-        f"SELECT COUNT(*) AS call_count FROM usage_events AS u {where_clause}",
-        params,
-    ).fetchone()
-    total_matched = int(count_row["call_count"] if count_row is not None else 0)
-    rows = conn.execute(
-        f"""
-        SELECT
-            u.record_id,
-            u.session_id,
-            u.thread_name,
-            u.parent_thread_name,
-            u.thread_key,
-            u.event_timestamp,
-            u.model,
-            u.effort,
-            u.total_tokens,
-            u.input_tokens,
-            u.cached_input_tokens,
-            u.uncached_input_tokens,
-            u.output_tokens,
-            u.reasoning_output_tokens,
-            u.is_archived,
-            cf.fragment_id,
-            cf.turn_key,
-            cf.fragment_kind,
-            cf.role,
-            cf.safe_label,
-            cf.content_hash,
-            cf.content_size_bytes,
-            cf.fragment_text,
-            cf.includes_raw_fragment,
-            cf.source_file_id,
-            cf.line_start,
-            cf.line_end,
-            cf.token_link_record_id
-        FROM (
-            SELECT *
-            FROM usage_events AS u
-            {where_clause}
-            ORDER BY u.event_timestamp ASC, u.cumulative_total_tokens ASC
-            {_limit_clause(normalized_limit)}
-        ) AS u
-        LEFT JOIN content_fragments AS cf ON cf.record_id = u.record_id
-        ORDER BY u.event_timestamp ASC, u.cumulative_total_tokens ASC, cf.line_start ASC
-        """,
-        [*params, *_limit_params(normalized_limit, normalized_offset)],
-    ).fetchall()
-    return ContentTraceResult(
-        calls=_trace_calls(rows, max_snippet_chars=max_snippet_chars),
-        total_matched_calls=total_matched,
-    )
-
-
-def _trace_where_clause(
-    *,
-    thread: str | None,
-    thread_key: str | None,
-    session_id: str | None,
-    record_id: str | None,
-    since: str | None,
-    until: str | None,
-    include_archived: bool,
-) -> tuple[str, list[object]]:
-    usage_clause, params = usage_where_clause(
-        since=since,
-        until=until,
-        table_alias="u",
-        include_archived=include_archived,
-    )
-    clauses: list[str] = []
-    if usage_clause:
-        clauses.append(usage_clause.removeprefix("WHERE "))
-    identity_clauses: list[str] = []
-    identity_params: list[object] = []
-    if thread:
-        identity_clauses.append(
-            "(u.thread_name = ? OR u.parent_thread_name = ? OR u.session_id = ?)"
-        )
-        identity_params.extend([thread, thread, thread])
-    if thread_key:
-        identity_clauses.append("u.thread_key = ?")
-        identity_params.append(thread_key)
-    if session_id:
-        identity_clauses.append("u.session_id = ?")
-        identity_params.append(session_id)
-    if record_id:
-        identity_clauses.append(
-            """
-            (
-                u.record_id = ?
-                OR u.session_id = (
-                    SELECT session_id FROM usage_events WHERE record_id = ?
-                )
-                OR (
-                    u.thread_key IS NOT NULL
-                    AND u.thread_key = (
-                        SELECT thread_key FROM usage_events WHERE record_id = ?
-                    )
-                )
-            )
-            """
-        )
-        identity_params.extend([record_id, record_id, record_id])
-    if not identity_clauses:
-        raise ValueError("thread, thread_key, session_id, or record_id is required")
-    clauses.append("(" + " OR ".join(identity_clauses) + ")")
-    return "WHERE " + " AND ".join(clauses), [*params, *identity_params]
-
-
-def _trace_calls(
-    rows: list[sqlite3.Row],
-    *,
-    max_snippet_chars: int | None,
-) -> list[dict[str, object]]:
-    calls_by_id: dict[str, dict[str, object]] = {}
-    for row in rows:
-        record_id = str(row["record_id"])
-        call = calls_by_id.setdefault(record_id, _trace_call_row(row))
-        if row["fragment_id"] is not None:
-            fragments = call["fragments"]
-            if not isinstance(fragments, list):
-                raise TypeError("trace fragments must be a list")
-            fragments.append(_trace_fragment_row(row, max_snippet_chars=max_snippet_chars))
-    for call in calls_by_id.values():
-        fragments = call["fragments"]
-        if not isinstance(fragments, list):
-            raise TypeError("trace fragments must be a list")
-        call["fragment_count"] = len(fragments)
-    return list(calls_by_id.values())
-
-
-def _trace_call_row(row: sqlite3.Row) -> dict[str, object]:
-    return {
-        "record_id": row["record_id"],
-        "session_id": row["session_id"],
-        "thread_name": row["thread_name"],
-        "parent_thread_name": row["parent_thread_name"],
-        "thread_key": row["thread_key"],
-        "event_timestamp": row["event_timestamp"],
-        "model": row["model"],
-        "effort": row["effort"],
-        "total_tokens": int(row["total_tokens"] or 0),
-        "input_tokens": int(row["input_tokens"] or 0),
-        "cached_input_tokens": int(row["cached_input_tokens"] or 0),
-        "uncached_input_tokens": int(row["uncached_input_tokens"] or 0),
-        "output_tokens": int(row["output_tokens"] or 0),
-        "reasoning_output_tokens": int(row["reasoning_output_tokens"] or 0),
-        "is_archived": bool(row["is_archived"]),
-        "fragment_count": 0,
-        "fragments": [],
-    }
-
-
-def _trace_fragment_row(
-    row: sqlite3.Row,
-    *,
-    max_snippet_chars: int | None,
-) -> dict[str, object]:
-    snippet, truncated = _snippet(
-        str(row["fragment_text"] or ""),
-        query="",
-        max_chars=max_snippet_chars,
-    )
-    return {
-        "fragment_id": row["fragment_id"],
-        "turn_key": row["turn_key"],
-        "fragment_kind": row["fragment_kind"],
-        "role": row["role"],
-        "safe_label": row["safe_label"],
-        "content_hash": row["content_hash"],
-        "content_size_bytes": int(row["content_size_bytes"] or 0),
-        "snippet": snippet,
-        "snippet_truncated": truncated,
-        "includes_raw_fragment": bool(row["includes_raw_fragment"]),
-        "source_file_id": row["source_file_id"],
-        "line_start": row["line_start"],
-        "line_end": row["line_end"],
-        "token_link_record_id": row["token_link_record_id"],
-    }
-
-
-def _search_content_fts(
-    conn: sqlite3.Connection,
-    *,
-    query: str,
-    since: str | None,
-    until: str | None,
-    model: str | None,
-    effort: str | None,
-    thread: str | None,
-    include_archived: bool,
-    limit: int | None,
-    offset: int,
-    max_snippet_chars: int | None,
-) -> ContentSearchResult:
-    usage_clause, usage_params = _search_usage_clause(
-        since=since,
-        until=until,
-        model=model,
-        effort=effort,
-        thread=thread,
-        include_archived=include_archived,
-    )
-    match_query = _fts_match_query(query)
-    from_sql = f"""
-        FROM content_fts
-        JOIN content_fragments AS cf ON cf.fragment_rowid = content_fts.rowid
-        JOIN usage_events AS u ON u.record_id = cf.record_id
-        WHERE content_fts MATCH ?
-        {usage_clause}
-    """
-    params: list[object] = [match_query, *usage_params]
-    total_matched = _search_count(conn, from_sql=from_sql, params=params)
-    rows = conn.execute(
-        f"""
-        SELECT
-            bm25(content_fts) AS search_rank,
-            cf.fragment_id,
-            cf.record_id,
-            cf.turn_key,
-            cf.fragment_kind,
-            cf.role,
-            cf.safe_label,
-            cf.content_hash,
-            cf.content_size_bytes,
-            cf.fragment_text,
-            cf.includes_raw_fragment,
-            cf.source_file_id,
-            cf.line_start,
-            cf.line_end,
-            cf.token_link_record_id,
-            u.session_id,
-            u.thread_name,
-            u.parent_thread_name,
-            u.thread_key,
-            u.event_timestamp,
-            u.model,
-            u.effort,
-            u.total_tokens,
-            u.input_tokens,
-            u.cached_input_tokens,
-            u.uncached_input_tokens,
-            u.output_tokens,
-            u.reasoning_output_tokens,
-            u.is_archived
-        {from_sql}
-        ORDER BY search_rank ASC, u.event_timestamp DESC, cf.line_start ASC
-        {_limit_clause(limit)}
-        """,
-        [*params, *_limit_params(limit, offset)],
-    ).fetchall()
-    return ContentSearchResult(
-        rows=[
-            _content_search_row(
-                row,
-                query=query,
-                max_snippet_chars=max_snippet_chars,
-            )
-            for row in rows
-        ],
-        total_matched_rows=total_matched,
-        search_mode="fts5",
-    )
-
-
-def _search_content_like(
-    conn: sqlite3.Connection,
-    *,
-    query: str,
-    since: str | None,
-    until: str | None,
-    model: str | None,
-    effort: str | None,
-    thread: str | None,
-    include_archived: bool,
-    limit: int | None,
-    offset: int,
-    max_snippet_chars: int | None,
-) -> ContentSearchResult:
-    usage_clause, usage_params = _search_usage_clause(
-        since=since,
-        until=until,
-        model=model,
-        effort=effort,
-        thread=thread,
-        include_archived=include_archived,
-    )
-    terms = _search_terms(query) or [query]
-    like_clauses = [
-        "(lower(cf.fragment_text) LIKE ? OR lower(cf.safe_label) LIKE ?)" for _term in terms
-    ]
-    like_params: list[object] = []
-    for term in terms:
-        pattern = f"%{term.lower()}%"
-        like_params.extend([pattern, pattern])
-    from_sql = f"""
-        FROM content_fragments AS cf
-        JOIN usage_events AS u ON u.record_id = cf.record_id
-        WHERE {" AND ".join(like_clauses)}
-        {usage_clause}
-    """
-    params = [*like_params, *usage_params]
-    total_matched = _search_count(conn, from_sql=from_sql, params=params)
-    rows = conn.execute(
-        f"""
-        SELECT
-            NULL AS search_rank,
-            cf.fragment_id,
-            cf.record_id,
-            cf.turn_key,
-            cf.fragment_kind,
-            cf.role,
-            cf.safe_label,
-            cf.content_hash,
-            cf.content_size_bytes,
-            cf.fragment_text,
-            cf.includes_raw_fragment,
-            cf.source_file_id,
-            cf.line_start,
-            cf.line_end,
-            cf.token_link_record_id,
-            u.session_id,
-            u.thread_name,
-            u.parent_thread_name,
-            u.thread_key,
-            u.event_timestamp,
-            u.model,
-            u.effort,
-            u.total_tokens,
-            u.input_tokens,
-            u.cached_input_tokens,
-            u.uncached_input_tokens,
-            u.output_tokens,
-            u.reasoning_output_tokens,
-            u.is_archived
-        {from_sql}
-        ORDER BY u.event_timestamp DESC, cf.line_start ASC
-        {_limit_clause(limit)}
-        """,
-        [*params, *_limit_params(limit, offset)],
-    ).fetchall()
-    return ContentSearchResult(
-        rows=[
-            _content_search_row(
-                row,
-                query=query,
-                max_snippet_chars=max_snippet_chars,
-            )
-            for row in rows
-        ],
-        total_matched_rows=total_matched,
-        search_mode="like",
-    )
-
-
-def _search_usage_clause(
-    *,
-    since: str | None,
-    until: str | None,
-    model: str | None,
-    effort: str | None,
-    thread: str | None,
-    include_archived: bool,
-) -> tuple[str, list[object]]:
-    where_clause, params = usage_where_clause(
-        since=since,
-        until=until,
-        model=model,
-        effort=effort,
-        thread=thread,
-        table_alias="u",
-        include_archived=include_archived,
-    )
-    if not where_clause:
-        return "", []
-    return f"AND {where_clause.removeprefix('WHERE ')}", params
-
-
-def _search_count(
-    conn: sqlite3.Connection,
-    *,
-    from_sql: str,
-    params: list[object],
-) -> int:
-    row = conn.execute(f"SELECT COUNT(*) AS row_count {from_sql}", params).fetchone()
-    return int(row["row_count"] if row is not None else 0)
-
-
-def _content_search_row(
-    row: sqlite3.Row,
-    *,
-    query: str,
-    max_snippet_chars: int | None,
-) -> dict[str, object]:
-    fragment_text = str(row["fragment_text"] or "")
-    snippet, truncated = _snippet(
-        fragment_text,
-        query=query,
-        max_chars=max_snippet_chars,
-    )
-    return {
-        "record_id": row["record_id"],
-        "session_id": row["session_id"],
-        "thread_name": row["thread_name"],
-        "parent_thread_name": row["parent_thread_name"],
-        "thread_key": row["thread_key"],
-        "event_timestamp": row["event_timestamp"],
-        "model": row["model"],
-        "effort": row["effort"],
-        "total_tokens": int(row["total_tokens"] or 0),
-        "input_tokens": int(row["input_tokens"] or 0),
-        "cached_input_tokens": int(row["cached_input_tokens"] or 0),
-        "uncached_input_tokens": int(row["uncached_input_tokens"] or 0),
-        "output_tokens": int(row["output_tokens"] or 0),
-        "reasoning_output_tokens": int(row["reasoning_output_tokens"] or 0),
-        "is_archived": bool(row["is_archived"]),
-        "fragment_id": row["fragment_id"],
-        "turn_key": row["turn_key"],
-        "fragment_kind": row["fragment_kind"],
-        "role": row["role"],
-        "safe_label": row["safe_label"],
-        "content_hash": row["content_hash"],
-        "content_size_bytes": int(row["content_size_bytes"] or 0),
-        "snippet": snippet,
-        "snippet_truncated": truncated,
-        "includes_raw_fragment": bool(row["includes_raw_fragment"]),
-        "source_file_id": row["source_file_id"],
-        "line_start": row["line_start"],
-        "line_end": row["line_end"],
-        "token_link_record_id": row["token_link_record_id"],
-        "search_rank": row["search_rank"],
-    }
-
-
-def _snippet(
-    text: str,
-    *,
-    query: str,
-    max_chars: int | None,
-) -> tuple[str, bool]:
-    if max_chars is None or max_chars <= 0 or len(text) <= max_chars:
-        return text, False
-    terms = _search_terms(query)
-    lower_text = text.lower()
-    positions = [lower_text.find(term.lower()) for term in terms]
-    match_positions = [position for position in positions if position >= 0]
-    center = min(match_positions) if match_positions else 0
-    start = max(0, center - max_chars // 3)
-    end = min(len(text), start + max_chars)
-    if end - start < max_chars:
-        start = max(0, end - max_chars)
-    excerpt = text[start:end].strip()
-    prefix = "... " if start > 0 else ""
-    suffix = " ..." if end < len(text) else ""
-    return f"{prefix}{excerpt}{suffix}", True
-
-
-def _search_terms(query: str) -> list[str]:
-    return [term for term in re.findall(r"[\w-]+", query) if term]
-
-
-def _fts_match_query(query: str) -> str:
-    terms = _search_terms(query)
-    if not terms:
-        return _fts_quote(query)
-    return " ".join(_fts_quote(term) for term in terms)
-
-
-def _fts_quote(value: str) -> str:
-    return '"' + value.replace('"', '""') + '"'
-
-
-def _normalize_search_limit(limit: int | None) -> int | None:
-    if limit is None or limit <= 0:
-        return None
-    return limit
-
-
-def _normalize_search_offset(offset: int) -> int:
-    return max(0, offset)
-
-
-def _limit_clause(limit: int | None) -> str:
-    if limit is None:
-        return ""
-    return "LIMIT ? OFFSET ?"
-
-
-def _limit_params(limit: int | None, offset: int) -> list[int]:
-    if limit is None:
-        return []
-    return [limit, offset]
 
 
 def _content_usage_rows_for_plans(
